move reminders into new service to avoid complicated test dependencies
