+++ backend/src/main/java/gov/cdc/usds/simplereport/SimpleReportApplication.java
@@ -63,7 +63,7 @@ public class SimpleReportApplication {
   @Bean
   @ConditionalOnProperty("simple-report.id-verification-reminders.enabled")
   public CommandLineRunner scheduleAccountReminderEmails(ScheduledTasksService scheduler) {
-    return args -> scheduler.scheduleAccountReminderEmails("0 * 1 * * *", "America/New_York");
+    return args -> scheduler.scheduleAccountReminderEmails("0 0 1 * * *", "America/New_York");
   }
 
   @Bean
